Move aktualizr_secondary loop to another class
